testiso: move qmp methods to a util file
move the qmp related methods to a separate file to make it cleaner
